feat(zint): support foundry tests
